speed_teste/requirements.txt
pillow==10.2.0
speedtest-cli==2.1.3
